+++ CHANGELOG.md
@@ -14,3 +14,8 @@
 - Raise min. Dart SDK to v3.0
 - Upgrade dependencies
 - Add new extension: `FileExtensions` to zip/unzip files
+
+## 0.1.1
+
+- Add possibility to auto-create a certificate if there's no valid one when doing project setup.
+- Fix bug were auto-select the first certificate of the profile and the profile has many certificates.